satellite/audit: mv ReservoirService into its own file (#2886)
