+++ satellite/audit/service.go
@@ -5,7 +5,6 @@ package audit
 
 import (
 	"context"
-	"math/rand"
 	"time"
 
 	"github.com/zeebo/errs"
@@ -59,50 +58,6 @@ func NewService(log *zap.Logger, config Config, metainfo *metainfo.Service,
 
 		Loop: *sync2.NewCycle(config.Interval),
 	}, nil
-}
-
-// ReservoirService is a temp name for the service struct during the audit 2.0 refactor.
-// Once V3-2363 and V3-2364 are implemented, Service2 will replace the existing Service struct.
-type ReservoirService struct {
-	log *zap.Logger
-
-	reservoirSlots int
-	Reservoirs     map[storj.NodeID]*Reservoir
-	rand           *rand.Rand
-
-	MetainfoLoop *metainfo.Loop
-	Loop         sync2.Cycle
-}
-
-// NewReservoirService instantiates Service2
-func NewReservoirService(log *zap.Logger, metaLoop *metainfo.Loop, config Config) (*ReservoirService, error) {
-	return &ReservoirService{
-		log: log,
-
-		reservoirSlots: config.Slots,
-		rand:           rand.New(rand.NewSource(time.Now().Unix())),
-
-		MetainfoLoop: metaLoop,
-		Loop:         *sync2.NewCycle(config.Interval),
-	}, nil
-}
-
-// Run runs auditing service 2.0
-func (service *ReservoirService) Run(ctx context.Context) (err error) {
-	defer mon.Task()(&ctx)(&err)
-	service.log.Info("audit 2.0 is starting up")
-
-	return service.Loop.Run(ctx, func(ctx context.Context) (err error) {
-		defer mon.Task()(&ctx)(&err)
-		pathCollector := NewPathCollector(service.reservoirSlots, service.rand)
-		err = service.MetainfoLoop.Join(ctx, pathCollector)
-		if err != nil {
-			service.log.Error("error joining metainfoloop", zap.Error(err))
-			return nil
-		}
-		service.Reservoirs = pathCollector.Reservoirs
-		return nil
-	})
 }
 
 // Run runs auditing service
